First pass at pulling nsoToken + Handler code
