+++ modules/splatinfo.py
