+++ modules/stringcrypt.py
